requirements.txt
flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
requests>=2.32.2  # Updated to resolve conflicts
python-dotenv==1.0.1
eventlet==0.35.2  # Required for Flask-SocketIO
gevent==23.9.1    # Downgraded for Python 3.10 compatibility
gevent-websocket==0.10.1

# FastAPI dependencies for multiprocess setup
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.25.0
pydantic>=2.0.0

# Game Boy Advance emulator dependencies
pygba==0.2.9      # For Game Boy emulation
mgba==0.10.2      # For GBA emulation

# Additional dependencies to resolve conflicts
pyyaml>=5.1       # Required by multiple packages
xxhash             # Required by datasets and evaluate
more-itertools     # Required by lm-eval
pytz>=2020.1      # Required by pandas
six>=1.14.0       # Required by rouge-score
colorama           # Required by sacrebleu
grpcio>=1.48.2    # Required by tensorboard
protobuf!=4.24.0,<5.0.0,>=3.19.6  # Required by tensorboard 

# VLM Backend Dependencies
openai==1.90.0
google-generativeai>=0.3.0  # Google Gemini API

# Local VLM Dependencies (HuggingFace + bitsandbytes)
torch>=2.0.0       # PyTorch for local models
transformers>=4.36.0  # HuggingFace transformers
bitsandbytes>=0.41.0  # Quantization support
accelerate>=0.25.0    # Model acceleration
sentencepiece>=0.1.99  # Tokenization support
protobuf>=3.20.0      # Protocol buffers